+++ modules/local/liana/rankaggregate/templates/rank_aggregate.py
@@ -131,8 +131,7 @@ if adata.obs[obs_key].nunique() > 1:
     )
     _log_subsample_info(subsample_info)
 
-    if (adata.X < 0).nnz == 0:
-        sc.pp.log1p(adata)
+    sc.pp.log1p(adata)
     try:
         li.mt.rank_aggregate(
             adata,